Bugfix: Report missing SNPs and 0 gene SNPs
Previously, if a SNP was missing from the --snp-intervals file, then it was
not reported in the snp_genes.txt output.

Likewise, if a SNP interval overlapped no genes, then it was not reported.

Also, clean up the log output.

Also, merge SNPs with the ',' character instead of the ':' character, since we
commonly have SNPs with names like '13:123456'.